requirements.txt
asgiref==3.3.4
Django==3.2.4
django-crispy-forms==1.12.0
gunicorn==20.1.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1